refactor: extract WMS operation partials from WmsRequestHandlers (#1144)
WmsRequestHandlers.cs was flagged as a god-file (2,653 LOC, 63 static
methods) in the audit-fixes review. Following the WFS split pattern,
move per-operation logic into partial-class files in the same folder:

- WmsRequestHandlers.GetMap.cs (298 LOC): HandleWmsGetMap
- WmsRequestHandlers.GetFeatureInfo.cs (282 LOC): HandleWmsGetFeatureInfo
- WmsRequestHandlers.GetCapabilities.cs (579 LOC): BuildWmsCapabilities,
  WMS dimension/online-resource/bbox helpers, and the
  CiteWmsDimensionDefinition record (only used by the capabilities
  builder)

The main file drops from 2,653 to 1,564 LOC. Public surface is
preserved; all helpers stay `private` (visible across partials in the
same assembly). The CITE GetMap helpers, validation helpers, and
WmsImageResult sealed type remain in the main file.

WMS test suite: 58 passed / 1 failed / 6 skipped before and after
(the failing MigrationScannerEndpointTests case is pre-existing and
unrelated).
