+++ src/Honua.Server/Features/Protocols/Ogc/Classic/Wms/WmsRequestHandlers.GetFeatureInfo.cs
@@ -1,0 +1,282 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+using System.Diagnostics;
+using System.Globalization;
+using System.Text;
+using Honua.Core.Features.Catalog.Domain;
+using Honua.Core.Features.FeatureStore.Abstractions;
+using Honua.Core.Features.FeatureStore.Domain;
+using Honua.Server.Features.Infrastructure.Helpers;
+using Honua.Server.Features.Infrastructure.Monitoring;
+using Honua.Server.Features.Infrastructure.Rendering;
+using Honua.Server.Features.Protocols.Ogc.Classic;
+using Honua.Server.Features.Protocols.Ogc.Common;
+using Honua.ServiceDefaults;
+using Microsoft.Extensions.DependencyInjection;
+using static Honua.Server.Features.Infrastructure.Rendering.RasterMapRenderingPipeline;
+using static Honua.Server.Features.Protocols.Ogc.Classic.OgcClassicRequestHelpers;
+
+namespace Honua.Server.Features.Protocols.Ogc.Classic.Wms;
+
+internal static partial class WmsRequestHandlers
+{
+    private static async Task<IResult> HandleWmsGetFeatureInfo(
+        HttpContext context,
+        ServiceDefinition service,
+        string serviceId,
+        ILogger logger)
+    {
+        var cancellationToken = TimeoutTokenHelper.GetTimeoutAwareCancellationToken(context);
+        OgcClassicLog.WmsRequested(logger, serviceId, "GetFeatureInfo");
+        using var activity = HonuaTelemetry.ActivitySource.StartActivity(
+            HonuaTelemetry.Activities.FeatureIdentify, ActivityKind.Internal);
+        activity?.SetTag(HonuaTelemetry.Tags.Protocol, HonuaTelemetry.Protocols.OgcMaps);
+        activity?.SetTag(HonuaTelemetry.Tags.ServiceId, serviceId);
+        activity?.SetTag(HonuaTelemetry.Tags.Operation, "wms-getfeatureinfo");
+
+        var query = context.Request.Query;
+
+        if (!TryGetRequiredQueryValue(query, "LAYERS", out var layersParam) ||
+            !TryParseCsvTokens(layersParam, allowEmptyTokens: false, out var mapLayerTokens))
+        {
+            return CreateWmsServiceException(context, "MissingParameterValue", "LAYERS parameter is required.");
+        }
+
+        if (!TryGetRequiredQueryValue(query, "QUERY_LAYERS", out var queryLayersParam) ||
+            !TryParseCsvTokens(queryLayersParam, allowEmptyTokens: false, out var queryLayerTokens))
+        {
+            return CreateWmsServiceException(context, "MissingParameterValue", "QUERY_LAYERS parameter is required.");
+        }
+
+        if (!TryGetRequiredQueryValue(query, "BBOX", out var bboxValue))
+        {
+            return CreateWmsServiceException(context, "InvalidParameterValue", "Invalid BBOX parameter. Expected format: xmin,ymin,xmax,ymax.");
+        }
+
+        if (!TryGetRequiredQueryValue(query, "VERSION", out var versionValue))
+        {
+            return CreateWmsServiceException(context, "MissingParameterValue", "VERSION parameter is required.");
+        }
+
+        if (!IsSupportedWmsVersion(versionValue))
+        {
+            return CreateWmsServiceException(context, "InvalidParameterValue", "Unsupported VERSION value. Supported values are 1.3.0 and 1.1.1.");
+        }
+
+        var crsValue = GetQueryValue(query, "CRS");
+        if (string.IsNullOrWhiteSpace(crsValue))
+        {
+            crsValue = GetQueryValue(query, "SRS");
+        }
+
+        if (!TryParseWmsCrs(crsValue, out var requestSrid, out var normalizedCrs))
+        {
+            return CreateWmsServiceException(context, "InvalidCRS", "Invalid or missing CRS/SRS parameter.");
+        }
+
+        if (!TryParseWmsBbox(bboxValue, normalizedCrs, versionValue, out var requestedExtent))
+        {
+            return CreateWmsServiceException(context, "InvalidParameterValue", "Invalid BBOX parameter. Expected format: xmin,ymin,xmax,ymax.");
+        }
+
+        if (!IsExtentWithinCrsBounds(requestedExtent, normalizedCrs))
+        {
+            return CreateWmsServiceException(context, "InvalidParameterValue", "BBOX is outside the valid range for the requested CRS.");
+        }
+
+        if (!TryGetRequiredQueryValue(query, "WIDTH", out var widthValue) ||
+            !int.TryParse(widthValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out var imageWidth) ||
+            imageWidth <= 0 || imageWidth > WmsMaxImageDimension)
+        {
+            return CreateWmsServiceException(context, "InvalidDimensionValue", $"WIDTH must be an integer between 1 and {WmsMaxImageDimension.ToString(CultureInfo.InvariantCulture)}.");
+        }
+
+        if (!TryGetRequiredQueryValue(query, "HEIGHT", out var heightValue) ||
+            !int.TryParse(heightValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out var imageHeight) ||
+            imageHeight <= 0 || imageHeight > WmsMaxImageDimension)
+        {
+            return CreateWmsServiceException(context, "InvalidDimensionValue", $"HEIGHT must be an integer between 1 and {WmsMaxImageDimension.ToString(CultureInfo.InvariantCulture)}.");
+        }
+
+        if (!TryParseWmsFeatureInfoPixel(query, versionValue, imageWidth, imageHeight, out var pixelX, out var pixelY))
+        {
+            var pointParameters = IsWms111Version(versionValue) ? "X/Y" : "I/J";
+            return CreateWmsServiceException(context, "InvalidPoint", $"{pointParameters} must be within the request image dimensions.");
+        }
+
+        if (!TryResolveWmsRequestedLayers(service, context, mapLayerTokens, out var mapLayers, out var unresolvedMapLayer))
+        {
+            var layerLabel = string.IsNullOrWhiteSpace(unresolvedMapLayer) ? "requested layer" : unresolvedMapLayer;
+            return CreateWmsServiceException(context, "LayerNotDefined", $"Layer '{layerLabel}' is not defined.");
+        }
+
+        if (!TryResolveWmsRequestedLayers(service, context, queryLayerTokens, out var queryLayers, out var unresolvedQueryLayer))
+        {
+            var layerLabel = string.IsNullOrWhiteSpace(unresolvedQueryLayer) ? "requested layer" : unresolvedQueryLayer;
+            return CreateWmsServiceException(context, "LayerNotDefined", $"Layer '{layerLabel}' is not defined.");
+        }
+
+        var mapLayerIds = new HashSet<int>(mapLayers.Select(layer => layer.Id));
+        foreach (var layer in queryLayers)
+        {
+            if (!mapLayerIds.Contains(layer.Id))
+            {
+                return CreateWmsServiceException(context, "LayerNotDefined", "QUERY_LAYERS must be a subset of LAYERS.");
+            }
+        }
+
+        if (!TryNormalizeFeatureInfoFormat(GetQueryValue(query, "INFO_FORMAT"), out var infoFormat))
+        {
+            return CreateWmsServiceException(context, "InvalidFormat", "Unsupported INFO_FORMAT. Supported values are text/plain and application/json.");
+        }
+
+        var filterResult = TryParseWmsLayerFilters(context, query, mapLayers);
+        if (filterResult.Error != null)
+        {
+            return filterResult.Error;
+        }
+        var filtersByLayerId = filterResult.Filters is null
+            ? null
+            : mapLayers
+                .Select((layer, index) => (layer.Id, Filter: filterResult.Filters[index]))
+                .Where(item => item.Filter is not null)
+                .ToDictionary(item => item.Id, item => item.Filter);
+
+        var featureCount = DefaultFeatureInfoCount;
+        var featureCountRaw = GetQueryValue(query, "FEATURE_COUNT");
+        if (!string.IsNullOrWhiteSpace(featureCountRaw))
+        {
+            if (!int.TryParse(featureCountRaw, NumberStyles.Integer, CultureInfo.InvariantCulture, out featureCount) || featureCount <= 0)
+            {
+                return CreateWmsServiceException(context, "InvalidParameterValue", "FEATURE_COUNT must be a positive integer.");
+            }
+        }
+
+        var queryExtent = requestedExtent;
+        if (requestSrid != service.SpatialReference.Srid)
+        {
+            var extentTransformResult = await TryTransformExtentAsync(
+                context,
+                requestedExtent,
+                requestSrid,
+                service.SpatialReference.Srid,
+                cancellationToken);
+            if (!extentTransformResult.IsSuccess)
+            {
+                return CreateWmsServiceException(context, "InvalidCRS", extentTransformResult.Error ?? "Invalid spatial reference.");
+            }
+
+            queryExtent = extentTransformResult.Extent;
+        }
+
+        var mapWidth = requestedExtent.MaxX - requestedExtent.MinX;
+        var mapHeight = requestedExtent.MaxY - requestedExtent.MinY;
+        var mapX = requestedExtent.MinX + (((pixelX + 0.5) / imageWidth) * mapWidth);
+        var mapY = requestedExtent.MaxY - (((pixelY + 0.5) / imageHeight) * mapHeight);
+
+        var toleranceX = Math.Max((mapWidth / imageWidth) * DefaultFeatureInfoTolerancePixels, 0.000001);
+        var toleranceY = Math.Max((mapHeight / imageHeight) * DefaultFeatureInfoTolerancePixels, 0.000001);
+        var clickExtent = new SkiaMapRenderer.RenderExtent(
+            mapX - toleranceX,
+            mapY - toleranceY,
+            mapX + toleranceX,
+            mapY + toleranceY);
+
+        if (requestSrid != service.SpatialReference.Srid)
+        {
+            var clickExtentTransform = await TryTransformExtentAsync(
+                context,
+                clickExtent,
+                requestSrid,
+                service.SpatialReference.Srid,
+                cancellationToken);
+            if (!clickExtentTransform.IsSuccess)
+            {
+                return CreateWmsServiceException(context, "InvalidCRS", clickExtentTransform.Error ?? "Invalid spatial reference.");
+            }
+
+            clickExtent = clickExtentTransform.Extent;
+        }
+
+        var featureReader = context.RequestServices.GetRequiredService<IFeatureReader>();
+        var spatialFilter = CreateBboxSpatialFilter(clickExtent, service.SpatialReference.Srid);
+        var remaining = Math.Min(featureCount, 1000);
+
+        var plainText = new StringBuilder();
+        var jsonFeatures = new List<WmsFeatureInfoFeature>();
+
+        foreach (var layer in queryLayers)
+        {
+            if (remaining <= 0)
+            {
+                break;
+            }
+
+            var featureQuery = new FeatureQuery
+            {
+                SpatialFilter = spatialFilter,
+                SqlFilter = filtersByLayerId != null && filtersByLayerId.TryGetValue(layer.Id, out var sqlFilter)
+                    ? sqlFilter
+                    : null,
+                SpatialReferenceSrid = service.SpatialReference.Srid,
+                OutputSrid = requestSrid,
+                Limit = remaining
+            };
+
+            var queryResult = await featureReader.QueryAsync(layer.Id, featureQuery, cancellationToken);
+            if (queryResult.Items.Length == 0)
+            {
+                continue;
+            }
+
+            foreach (var item in queryResult.Items)
+            {
+                if (remaining <= 0)
+                {
+                    break;
+                }
+
+                remaining--;
+                var layerName = GetWmsLayerName(layer);
+                var attributes = BuildVisibleFeatureInfoAttributes(item);
+
+                if (string.Equals(infoFormat, JsonMimeType, StringComparison.OrdinalIgnoreCase))
+                {
+                    jsonFeatures.Add(new WmsFeatureInfoFeature
+                    {
+                        Layer = layerName,
+                        Attributes = attributes
+                    });
+                    continue;
+                }
+
+                plainText.Append("Layer=").Append(layerName).AppendLine();
+                foreach (var attribute in attributes.OrderBy(pair => pair.Key, StringComparer.OrdinalIgnoreCase))
+                {
+                    plainText.Append(attribute.Key)
+                        .Append('=')
+                        .Append(FormatFeatureInfoValue(attribute.Value))
+                        .AppendLine();
+                }
+
+                plainText.AppendLine();
+            }
+        }
+
+        if (string.Equals(infoFormat, JsonMimeType, StringComparison.OrdinalIgnoreCase))
+        {
+            var payload = new WmsFeatureInfoResponse
+            {
+                Features = [.. jsonFeatures]
+            };
+
+            return Results.Json(payload, OgcClassicJsonContext.Default.WmsFeatureInfoResponse, contentType: JsonMimeType);
+        }
+
+        var body = plainText.Length > 0
+            ? plainText.ToString().TrimEnd()
+            : "No features found.";
+        return Results.Content(body, PlainTextMimeType, Encoding.UTF8, StatusCodes.Status200OK);
+    }
+}
