+++ src/Honua.Server/Features/Protocols/Ogc/Classic/Wms/WmsRequestHandlers.cs
@@ -30,7 +30,7 @@ using static Honua.Server.Features.Protocols.Ogc.Classic.OgcClassicRequestHelper
 
 namespace Honua.Server.Features.Protocols.Ogc.Classic.Wms;
 
-internal static class WmsRequestHandlers
+internal static partial class WmsRequestHandlers
 {
     private const int WmsMaxImageDimension = 4096;
     private const string Wms13Version = "1.3.0";
@@ -221,538 +221,6 @@ internal static class WmsRequestHandlers
             OgcClassicLog.WmsFailed(logger, serviceId, ex.Message, ex);
             return CreateWmsServiceException(context, "NoApplicableCode", "WMS request failed.", StatusCodes.Status500InternalServerError);
         }
-    }
-
-    private static async Task<IResult> HandleWmsGetMap(
-        HttpContext context,
-        ServiceDefinition service,
-        string serviceId,
-        ILogger logger)
-    {
-        var cancellationToken = TimeoutTokenHelper.GetTimeoutAwareCancellationToken(context);
-        OgcClassicLog.WmsRequested(logger, serviceId, "GetMap");
-        var stopwatch = Stopwatch.StartNew();
-        using var activity = HonuaTelemetry.ActivitySource.StartActivity(
-            HonuaTelemetry.Activities.MapRender, ActivityKind.Internal);
-        activity?.SetTag(HonuaTelemetry.Tags.Protocol, HonuaTelemetry.Protocols.OgcMaps);
-        activity?.SetTag(HonuaTelemetry.Tags.ServiceId, serviceId);
-        activity?.SetTag(HonuaTelemetry.Tags.Operation, "wms-getmap");
-
-        var query = context.Request.Query;
-        if (!TryGetRequiredQueryValue(query, "LAYERS", out var layersParam))
-        {
-            return CreateWmsServiceException(context, "MissingParameterValue", "LAYERS parameter is required.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "STYLES", out var stylesParam, allowEmpty: true))
-        {
-            return CreateWmsServiceException(context, "MissingParameterValue", "STYLES parameter is required.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "BBOX", out var bboxValue))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "Invalid BBOX parameter. Expected format: xmin,ymin,xmax,ymax.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "VERSION", out var versionValue))
-        {
-            return CreateWmsServiceException(context, "MissingParameterValue", "VERSION parameter is required.");
-        }
-
-        if (!IsSupportedWmsVersion(versionValue))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "Unsupported VERSION value. Supported values are 1.3.0 and 1.1.1.");
-        }
-
-        var crsValue = GetQueryValue(query, "CRS");
-        if (string.IsNullOrWhiteSpace(crsValue))
-        {
-            crsValue = GetQueryValue(query, "SRS");
-        }
-
-        if (!TryParseWmsCrs(crsValue, out var requestSrid, out var normalizedCrs))
-        {
-            return CreateWmsServiceException(context, "InvalidCRS", "Invalid or missing CRS/SRS parameter.");
-        }
-
-        if (!TryParseWmsBbox(bboxValue, normalizedCrs, versionValue, out var requestedExtent))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "Invalid BBOX parameter. Expected format: xmin,ymin,xmax,ymax.");
-        }
-
-        var bboxOutsideCrsBounds = !IsExtentWithinCrsBounds(requestedExtent, normalizedCrs);
-
-        if (!TryGetRequiredQueryValue(query, "WIDTH", out var widthValue) ||
-            !int.TryParse(widthValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out var imageWidth) ||
-            imageWidth <= 0 || imageWidth > WmsMaxImageDimension)
-        {
-            return CreateWmsServiceException(context, "InvalidDimensionValue", $"WIDTH must be an integer between 1 and {WmsMaxImageDimension.ToString(CultureInfo.InvariantCulture)}.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "HEIGHT", out var heightValue) ||
-            !int.TryParse(heightValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out var imageHeight) ||
-            imageHeight <= 0 || imageHeight > WmsMaxImageDimension)
-        {
-            return CreateWmsServiceException(context, "InvalidDimensionValue", $"HEIGHT must be an integer between 1 and {WmsMaxImageDimension.ToString(CultureInfo.InvariantCulture)}.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "FORMAT", out var formatValue))
-        {
-            return CreateWmsServiceException(context, "MissingParameterValue", "FORMAT parameter is required.");
-        }
-
-        if (!TryNormalizeWmsMapFormat(formatValue, out var imageFormat, out var contentType))
-        {
-            return CreateWmsServiceException(context, "InvalidFormat", "FORMAT must be image/png or image/jpeg.");
-        }
-
-        var exceptionsValue = GetQueryValue(query, "EXCEPTIONS");
-        if (!IsSupportedWmsExceptionFormat(exceptionsValue))
-        {
-            return CreateWmsServiceException(context, "InvalidFormat", "Unsupported EXCEPTIONS format. Only XML exceptions are supported.");
-        }
-
-        if (!TryParseWmsTransparent(GetQueryValue(query, "TRANSPARENT"), out var transparent))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "TRANSPARENT must be TRUE or FALSE.");
-        }
-
-        var backgroundColor = SKColors.White;
-        var backgroundValue = GetQueryValue(query, "BGCOLOR");
-        if (!string.IsNullOrWhiteSpace(backgroundValue) &&
-            !TryParseWmsBackgroundColor(backgroundValue, out backgroundColor))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "BGCOLOR must be formatted as 0xRRGGBB or #RRGGBB.");
-        }
-
-        if (!TryParseCsvTokens(layersParam, allowEmptyTokens: false, out var layerTokens))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "LAYERS must contain at least one layer name.");
-        }
-
-        if (!TryResolveWmsRequestedLayers(service, context, layerTokens, out var renderLayers, out var unresolvedLayer))
-        {
-            var layerLabel = string.IsNullOrWhiteSpace(unresolvedLayer) ? "requested layer" : unresolvedLayer;
-            return CreateWmsServiceException(context, "LayerNotDefined", $"Layer '{layerLabel}' is not defined.");
-        }
-
-        if (!ValidateWmsStyles(stylesParam, renderLayers.Length, out var styleError))
-        {
-            return CreateWmsServiceException(context, "StyleNotDefined", styleError);
-        }
-
-        var filterResult = TryParseWmsLayerFilters(context, query, renderLayers);
-        if (filterResult.Error != null)
-        {
-            return filterResult.Error;
-        }
-        var layerFilters = filterResult.Filters;
-        activity?.SetTag("wms.filter_applied", layerFilters is not null);
-
-        var temporalResult = TryParseWmsLayerTemporalFilters(context, query, renderLayers);
-        if (temporalResult.Error != null)
-        {
-            return temporalResult.Error;
-        }
-        var layerTemporalFilters = temporalResult.Filters;
-        activity?.SetTag("wms.time_applied", layerTemporalFilters is not null);
-
-        var effectiveTransparent = transparent && string.Equals(imageFormat, "png", StringComparison.OrdinalIgnoreCase);
-        await using var renderLease = await context.RequestServices
-            .GetRequiredService<RasterRenderCapacityLimiter>()
-            .TryAcquireAsync(imageWidth, imageHeight, cancellationToken)
-            .ConfigureAwait(false);
-        if (renderLease is null)
-        {
-            return CreateWmsServiceException(context,
-                "NoApplicableCode",
-                RasterRenderCapacityLimiter.CapacityExceededMessage,
-                StatusCodes.Status503ServiceUnavailable);
-        }
-
-        if (bboxOutsideCrsBounds)
-        {
-            using var outsideSurface = SKSurface.Create(new SKImageInfo(imageWidth, imageHeight, SKColorType.Rgba8888, SKAlphaType.Premul));
-            if (outsideSurface is null)
-            {
-                return CreateWmsServiceException(context, "NoApplicableCode", "Failed to allocate render surface.", StatusCodes.Status500InternalServerError);
-            }
-
-            outsideSurface.Canvas.Clear(effectiveTransparent ? SKColors.Transparent : backgroundColor);
-            var outsideBytes = SkiaMapRenderer.EncodeSurface(outsideSurface, imageFormat);
-            return Results.Bytes(outsideBytes, contentType);
-        }
-
-        if (TryHandleCiteWmsGetMap(
-                context,
-                service,
-                renderLayers,
-                query,
-                requestedExtent,
-                imageWidth,
-                imageHeight,
-                imageFormat,
-                contentType,
-                effectiveTransparent,
-                backgroundColor,
-                layerFilters,
-                out var citeResult))
-        {
-            return citeResult;
-        }
-
-        var queryExtent = requestedExtent;
-        if (requestSrid != service.SpatialReference.Srid)
-        {
-            var extentTransformResult = await TryTransformExtentAsync(
-                context,
-                requestedExtent,
-                requestSrid,
-                service.SpatialReference.Srid,
-                cancellationToken);
-            if (!extentTransformResult.IsSuccess)
-            {
-                return CreateWmsServiceException(context, "InvalidCRS", extentTransformResult.Error ?? "Invalid spatial reference.");
-            }
-
-            queryExtent = extentTransformResult.Extent;
-        }
-
-        var spatialFilter = CreateBboxSpatialFilter(queryExtent, service.SpatialReference.Srid);
-
-        using var surface = SKSurface.Create(new SKImageInfo(imageWidth, imageHeight, SKColorType.Rgba8888, SKAlphaType.Premul));
-        if (surface is null)
-        {
-            return CreateWmsServiceException(context, "NoApplicableCode", "Failed to allocate render surface.", StatusCodes.Status500InternalServerError);
-        }
-
-        var featureReader = context.RequestServices.GetRequiredService<IFeatureReader>();
-        var styleCatalog = context.RequestServices.GetRequiredService<ILayerStyleCatalog>();
-        var mapConfig = service.Metadata?.MapServer;
-        var maxFeatures = mapConfig?.MaxFeaturesPerLayer ?? MaxFeaturesPerLayer;
-        var totalFeatureCount = 0;
-
-        var canvas = surface.Canvas;
-        canvas.Clear(effectiveTransparent ? SKColors.Transparent : backgroundColor);
-        var transformFn = SkiaMapRenderer.BuildTransform(requestedExtent, imageWidth, imageHeight);
-
-        for (var i = 0; i < renderLayers.Length; i++)
-        {
-            var layer = renderLayers[i];
-            cancellationToken.ThrowIfCancellationRequested();
-
-            if (!layer.HasGeometry)
-            {
-                continue;
-            }
-
-            var stylePlan = await GetRasterStylePlanAsync(
-                styleCatalog,
-                layer.Id,
-                cancellationToken).ConfigureAwait(false);
-            var featureQuery = CreateRasterFeatureQuery(
-                stylePlan,
-                spatialFilter,
-                service.SpatialReference.Srid,
-                requestSrid,
-                maxFeatures,
-                layerFilters?[i],
-                layerTemporalFilters?[i]);
-
-            var renderedPointCount = await TryRenderRasterPointFastPathAsync(
-                canvas,
-                featureReader,
-                layer.Id,
-                layer.GeometryType,
-                stylePlan,
-                featureQuery,
-                requestedExtent,
-                imageWidth,
-                imageHeight,
-                transformFn,
-                cancellationToken).ConfigureAwait(false);
-            if (renderedPointCount >= 0)
-            {
-                totalFeatureCount += renderedPointCount;
-                continue;
-            }
-
-            var features = await QueryRasterFeaturesAsync(featureReader, layer.Id, featureQuery, cancellationToken);
-            if (features.Length == 0)
-            {
-                continue;
-            }
-
-            totalFeatureCount += features.Length;
-            RenderLayerToCanvas(canvas, features, stylePlan.StyleLayers, transformFn, layer.GeometryType);
-        }
-
-        var imageBytes = SkiaMapRenderer.EncodeSurface(surface, imageFormat);
-
-        stopwatch.Stop();
-        HonuaTelemetry.SetSuccess(activity, totalFeatureCount);
-        HonuaTelemetry.CategorizeLatency(activity, stopwatch.Elapsed.TotalMilliseconds);
-
-        return Results.Bytes(imageBytes, contentType);
-    }
-
-    private static async Task<IResult> HandleWmsGetFeatureInfo(
-        HttpContext context,
-        ServiceDefinition service,
-        string serviceId,
-        ILogger logger)
-    {
-        var cancellationToken = TimeoutTokenHelper.GetTimeoutAwareCancellationToken(context);
-        OgcClassicLog.WmsRequested(logger, serviceId, "GetFeatureInfo");
-        using var activity = HonuaTelemetry.ActivitySource.StartActivity(
-            HonuaTelemetry.Activities.FeatureIdentify, ActivityKind.Internal);
-        activity?.SetTag(HonuaTelemetry.Tags.Protocol, HonuaTelemetry.Protocols.OgcMaps);
-        activity?.SetTag(HonuaTelemetry.Tags.ServiceId, serviceId);
-        activity?.SetTag(HonuaTelemetry.Tags.Operation, "wms-getfeatureinfo");
-
-        var query = context.Request.Query;
-
-        if (!TryGetRequiredQueryValue(query, "LAYERS", out var layersParam) ||
-            !TryParseCsvTokens(layersParam, allowEmptyTokens: false, out var mapLayerTokens))
-        {
-            return CreateWmsServiceException(context, "MissingParameterValue", "LAYERS parameter is required.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "QUERY_LAYERS", out var queryLayersParam) ||
-            !TryParseCsvTokens(queryLayersParam, allowEmptyTokens: false, out var queryLayerTokens))
-        {
-            return CreateWmsServiceException(context, "MissingParameterValue", "QUERY_LAYERS parameter is required.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "BBOX", out var bboxValue))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "Invalid BBOX parameter. Expected format: xmin,ymin,xmax,ymax.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "VERSION", out var versionValue))
-        {
-            return CreateWmsServiceException(context, "MissingParameterValue", "VERSION parameter is required.");
-        }
-
-        if (!IsSupportedWmsVersion(versionValue))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "Unsupported VERSION value. Supported values are 1.3.0 and 1.1.1.");
-        }
-
-        var crsValue = GetQueryValue(query, "CRS");
-        if (string.IsNullOrWhiteSpace(crsValue))
-        {
-            crsValue = GetQueryValue(query, "SRS");
-        }
-
-        if (!TryParseWmsCrs(crsValue, out var requestSrid, out var normalizedCrs))
-        {
-            return CreateWmsServiceException(context, "InvalidCRS", "Invalid or missing CRS/SRS parameter.");
-        }
-
-        if (!TryParseWmsBbox(bboxValue, normalizedCrs, versionValue, out var requestedExtent))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "Invalid BBOX parameter. Expected format: xmin,ymin,xmax,ymax.");
-        }
-
-        if (!IsExtentWithinCrsBounds(requestedExtent, normalizedCrs))
-        {
-            return CreateWmsServiceException(context, "InvalidParameterValue", "BBOX is outside the valid range for the requested CRS.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "WIDTH", out var widthValue) ||
-            !int.TryParse(widthValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out var imageWidth) ||
-            imageWidth <= 0 || imageWidth > WmsMaxImageDimension)
-        {
-            return CreateWmsServiceException(context, "InvalidDimensionValue", $"WIDTH must be an integer between 1 and {WmsMaxImageDimension.ToString(CultureInfo.InvariantCulture)}.");
-        }
-
-        if (!TryGetRequiredQueryValue(query, "HEIGHT", out var heightValue) ||
-            !int.TryParse(heightValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out var imageHeight) ||
-            imageHeight <= 0 || imageHeight > WmsMaxImageDimension)
-        {
-            return CreateWmsServiceException(context, "InvalidDimensionValue", $"HEIGHT must be an integer between 1 and {WmsMaxImageDimension.ToString(CultureInfo.InvariantCulture)}.");
-        }
-
-        if (!TryParseWmsFeatureInfoPixel(query, versionValue, imageWidth, imageHeight, out var pixelX, out var pixelY))
-        {
-            var pointParameters = IsWms111Version(versionValue) ? "X/Y" : "I/J";
-            return CreateWmsServiceException(context, "InvalidPoint", $"{pointParameters} must be within the request image dimensions.");
-        }
-
-        if (!TryResolveWmsRequestedLayers(service, context, mapLayerTokens, out var mapLayers, out var unresolvedMapLayer))
-        {
-            var layerLabel = string.IsNullOrWhiteSpace(unresolvedMapLayer) ? "requested layer" : unresolvedMapLayer;
-            return CreateWmsServiceException(context, "LayerNotDefined", $"Layer '{layerLabel}' is not defined.");
-        }
-
-        if (!TryResolveWmsRequestedLayers(service, context, queryLayerTokens, out var queryLayers, out var unresolvedQueryLayer))
-        {
-            var layerLabel = string.IsNullOrWhiteSpace(unresolvedQueryLayer) ? "requested layer" : unresolvedQueryLayer;
-            return CreateWmsServiceException(context, "LayerNotDefined", $"Layer '{layerLabel}' is not defined.");
-        }
-
-        var mapLayerIds = new HashSet<int>(mapLayers.Select(layer => layer.Id));
-        foreach (var layer in queryLayers)
-        {
-            if (!mapLayerIds.Contains(layer.Id))
-            {
-                return CreateWmsServiceException(context, "LayerNotDefined", "QUERY_LAYERS must be a subset of LAYERS.");
-            }
-        }
-
-        if (!TryNormalizeFeatureInfoFormat(GetQueryValue(query, "INFO_FORMAT"), out var infoFormat))
-        {
-            return CreateWmsServiceException(context, "InvalidFormat", "Unsupported INFO_FORMAT. Supported values are text/plain and application/json.");
-        }
-
-        var filterResult = TryParseWmsLayerFilters(context, query, mapLayers);
-        if (filterResult.Error != null)
-        {
-            return filterResult.Error;
-        }
-        var filtersByLayerId = filterResult.Filters is null
-            ? null
-            : mapLayers
-                .Select((layer, index) => (layer.Id, Filter: filterResult.Filters[index]))
-                .Where(item => item.Filter is not null)
-                .ToDictionary(item => item.Id, item => item.Filter);
-
-        var featureCount = DefaultFeatureInfoCount;
-        var featureCountRaw = GetQueryValue(query, "FEATURE_COUNT");
-        if (!string.IsNullOrWhiteSpace(featureCountRaw))
-        {
-            if (!int.TryParse(featureCountRaw, NumberStyles.Integer, CultureInfo.InvariantCulture, out featureCount) || featureCount <= 0)
-            {
-                return CreateWmsServiceException(context, "InvalidParameterValue", "FEATURE_COUNT must be a positive integer.");
-            }
-        }
-
-        var queryExtent = requestedExtent;
-        if (requestSrid != service.SpatialReference.Srid)
-        {
-            var extentTransformResult = await TryTransformExtentAsync(
-                context,
-                requestedExtent,
-                requestSrid,
-                service.SpatialReference.Srid,
-                cancellationToken);
-            if (!extentTransformResult.IsSuccess)
-            {
-                return CreateWmsServiceException(context, "InvalidCRS", extentTransformResult.Error ?? "Invalid spatial reference.");
-            }
-
-            queryExtent = extentTransformResult.Extent;
-        }
-
-        var mapWidth = requestedExtent.MaxX - requestedExtent.MinX;
-        var mapHeight = requestedExtent.MaxY - requestedExtent.MinY;
-        var mapX = requestedExtent.MinX + (((pixelX + 0.5) / imageWidth) * mapWidth);
-        var mapY = requestedExtent.MaxY - (((pixelY + 0.5) / imageHeight) * mapHeight);
-
-        var toleranceX = Math.Max((mapWidth / imageWidth) * DefaultFeatureInfoTolerancePixels, 0.000001);
-        var toleranceY = Math.Max((mapHeight / imageHeight) * DefaultFeatureInfoTolerancePixels, 0.000001);
-        var clickExtent = new SkiaMapRenderer.RenderExtent(
-            mapX - toleranceX,
-            mapY - toleranceY,
-            mapX + toleranceX,
-            mapY + toleranceY);
-
-        if (requestSrid != service.SpatialReference.Srid)
-        {
-            var clickExtentTransform = await TryTransformExtentAsync(
-                context,
-                clickExtent,
-                requestSrid,
-                service.SpatialReference.Srid,
-                cancellationToken);
-            if (!clickExtentTransform.IsSuccess)
-            {
-                return CreateWmsServiceException(context, "InvalidCRS", clickExtentTransform.Error ?? "Invalid spatial reference.");
-            }
-
-            clickExtent = clickExtentTransform.Extent;
-        }
-
-        var featureReader = context.RequestServices.GetRequiredService<IFeatureReader>();
-        var spatialFilter = CreateBboxSpatialFilter(clickExtent, service.SpatialReference.Srid);
-        var remaining = Math.Min(featureCount, 1000);
-
-        var plainText = new StringBuilder();
-        var jsonFeatures = new List<WmsFeatureInfoFeature>();
-
-        foreach (var layer in queryLayers)
-        {
-            if (remaining <= 0)
-            {
-                break;
-            }
-
-            var featureQuery = new FeatureQuery
-            {
-                SpatialFilter = spatialFilter,
-                SqlFilter = filtersByLayerId != null && filtersByLayerId.TryGetValue(layer.Id, out var sqlFilter)
-                    ? sqlFilter
-                    : null,
-                SpatialReferenceSrid = service.SpatialReference.Srid,
-                OutputSrid = requestSrid,
-                Limit = remaining
-            };
-
-            var queryResult = await featureReader.QueryAsync(layer.Id, featureQuery, cancellationToken);
-            if (queryResult.Items.Length == 0)
-            {
-                continue;
-            }
-
-            foreach (var item in queryResult.Items)
-            {
-                if (remaining <= 0)
-                {
-                    break;
-                }
-
-                remaining--;
-                var layerName = GetWmsLayerName(layer);
-                var attributes = BuildVisibleFeatureInfoAttributes(item);
-
-                if (string.Equals(infoFormat, JsonMimeType, StringComparison.OrdinalIgnoreCase))
-                {
-                    jsonFeatures.Add(new WmsFeatureInfoFeature
-                    {
-                        Layer = layerName,
-                        Attributes = attributes
-                    });
-                    continue;
-                }
-
-                plainText.Append("Layer=").Append(layerName).AppendLine();
-                foreach (var attribute in attributes.OrderBy(pair => pair.Key, StringComparer.OrdinalIgnoreCase))
-                {
-                    plainText.Append(attribute.Key)
-                        .Append('=')
-                        .Append(FormatFeatureInfoValue(attribute.Value))
-                        .AppendLine();
-                }
-
-                plainText.AppendLine();
-            }
-        }
-
-        if (string.Equals(infoFormat, JsonMimeType, StringComparison.OrdinalIgnoreCase))
-        {
-            var payload = new WmsFeatureInfoResponse
-            {
-                Features = [.. jsonFeatures]
-            };
-
-            return Results.Json(payload, OgcClassicJsonContext.Default.WmsFeatureInfoResponse, contentType: JsonMimeType);
-        }
-
-        var body = plainText.Length > 0
-            ? plainText.ToString().TrimEnd()
-            : "No features found.";
-        return Results.Content(body, PlainTextMimeType, Encoding.UTF8, StatusCodes.Status200OK);
     }
 
     private static bool TryResolveWmsRequestedLayers(
@@ -2065,553 +1533,6 @@ internal static class WmsRequestHandlers
         return new WmsImageResult(imageBytes, contentType, warningHeader);
     }
 
-    private static void AppendWmsCiteDimensions(StringBuilder sb, LayerDefinition layer, string indent, bool isWms111)
-    {
-        var definition = GetCiteWmsDimensionDefinition(layer);
-        if (definition is null)
-        {
-            return;
-        }
-
-        if (isWms111)
-        {
-            sb.Append(indent)
-                .Append("<Dimension name=\"")
-                .Append(EscapeXml(definition.Value.Name))
-                .Append("\" units=\"")
-                .Append(EscapeXml(definition.Value.Units))
-                .Append('"');
-
-            if (!string.IsNullOrWhiteSpace(definition.Value.UnitSymbol))
-            {
-                sb.Append(" unitSymbol=\"")
-                    .Append(EscapeXml(definition.Value.UnitSymbol!))
-                    .Append('"');
-            }
-
-            sb.AppendLine(" />");
-            sb.Append(indent)
-                .Append("<Extent name=\"")
-                .Append(EscapeXml(definition.Value.Name))
-                .Append('"');
-
-            if (!string.IsNullOrWhiteSpace(definition.Value.Default))
-            {
-                sb.Append(" default=\"")
-                    .Append(EscapeXml(definition.Value.Default!))
-                    .Append('"');
-            }
-
-            sb.Append(" nearestValue=\"")
-                .Append(definition.Value.NearestValue ? "1" : "0")
-                .Append("\">")
-                .Append(EscapeXml(definition.Value.Extent))
-                .AppendLine("</Extent>");
-            return;
-        }
-
-        sb.Append(indent)
-            .Append("<Dimension name=\"")
-            .Append(EscapeXml(definition.Value.Name))
-            .Append("\" units=\"")
-            .Append(EscapeXml(definition.Value.Units))
-            .Append('"');
-
-        if (!string.IsNullOrWhiteSpace(definition.Value.UnitSymbol))
-        {
-            sb.Append(" unitSymbol=\"")
-                .Append(EscapeXml(definition.Value.UnitSymbol!))
-                .Append('"');
-        }
-
-        sb.Append(" multipleValues=\"")
-            .Append(definition.Value.MultipleValues ? "true" : "false")
-            .Append("\" nearestValue=\"")
-            .Append(definition.Value.NearestValue ? "true" : "false")
-            .Append('"');
-
-        if (!string.IsNullOrWhiteSpace(definition.Value.Default))
-        {
-            sb.Append(" default=\"")
-                .Append(EscapeXml(definition.Value.Default!))
-                .Append('"');
-        }
-
-        if (!string.IsNullOrWhiteSpace(definition.Value.Current))
-        {
-            sb.Append(" current=\"")
-                .Append(EscapeXml(definition.Value.Current!))
-                .Append('"');
-        }
-
-        sb.Append('>')
-            .Append(EscapeXml(definition.Value.Extent))
-            .AppendLine("</Dimension>");
-    }
-
-    private static async Task AppendWmsTemporalDimensionAsync(
-        HttpContext context,
-        StringBuilder sb,
-        LayerDefinition layer,
-        string indent,
-        bool isWms111)
-    {
-        // CITE Autos has its own hardcoded "time" dimension already emitted by
-        // AppendWmsCiteDimensions; do not duplicate.
-        if (string.Equals(layer.Name, CiteAutosLayerTitle, StringComparison.OrdinalIgnoreCase))
-        {
-            return;
-        }
-
-        if (layer.Metadata?.TimeInfo is null ||
-            string.IsNullOrWhiteSpace(layer.Metadata.TimeInfo.StartTimeField))
-        {
-            return;
-        }
-
-        var featureReader = context.RequestServices.GetService<IFeatureReader>();
-        if (featureReader is null)
-        {
-            return;
-        }
-
-        var cancellationToken = TimeoutTokenHelper.GetTimeoutAwareCancellationToken(context);
-        var range = await TemporalExtentHelpers.TryResolveTemporalRangeAsync(
-            layer,
-            featureReader,
-            cancellationToken).ConfigureAwait(false);
-        if (range is null || !range.Value.HasExtent || range.Value.Min is null || range.Value.Max is null)
-        {
-            return;
-        }
-
-        var min = FormatWmsTemporalInstant(range.Value.Min.Value);
-        var max = FormatWmsTemporalInstant(range.Value.Max.Value);
-        var extent = $"{min}/{max}/PT0S";
-
-        if (isWms111)
-        {
-            sb.Append(indent)
-                .Append("<Dimension name=\"time\" units=\"ISO8601\" />")
-                .AppendLine();
-            sb.Append(indent)
-                .Append("<Extent name=\"time\" default=\"")
-                .Append(EscapeXml(max))
-                .Append("\" nearestValue=\"1\">")
-                .Append(EscapeXml(extent))
-                .AppendLine("</Extent>");
-            return;
-        }
-
-        sb.Append(indent)
-            .Append("<Dimension name=\"time\" units=\"ISO8601\" multipleValues=\"false\" nearestValue=\"true\" default=\"")
-            .Append(EscapeXml(max))
-            .Append("\">")
-            .Append(EscapeXml(extent))
-            .AppendLine("</Dimension>");
-    }
-
-    private static string FormatWmsTemporalInstant(DateTimeOffset value)
-        => TemporalExtentHelpers.FormatOgcTemporalValue(value);
-
-    private static void AppendWmsOnlineResource(StringBuilder sb, string indent, string href, bool isWms111)
-    {
-        sb.Append(indent).Append("<OnlineResource ");
-        if (isWms111)
-        {
-            sb.Append("xmlns:xlink=\"http://www.w3.org/1999/xlink\" xlink:type=\"simple\" ");
-        }
-
-        sb.Append("xlink:href=\"")
-            .Append(EscapeXml(href))
-            .AppendLine("\" />");
-    }
-
-    private static CiteWmsDimensionDefinition? GetCiteWmsDimensionDefinition(LayerDefinition layer)
-    {
-        if (string.Equals(layer.Name, CiteTerrainLayerTitle, StringComparison.OrdinalIgnoreCase))
-        {
-            return new CiteWmsDimensionDefinition(
-                Name: "elevation",
-                Units: "CRS:88",
-                UnitSymbol: "m",
-                MultipleValues: true,
-                NearestValue: false,
-                Default: CiteTerrainDefaultElevation,
-                Current: null,
-                Extent: "0/425/1");
-        }
-
-        if (string.Equals(layer.Name, CiteLakesLayerTitle, StringComparison.OrdinalIgnoreCase))
-        {
-            return new CiteWmsDimensionDefinition(
-                Name: "elevation",
-                Units: "CRS:88",
-                UnitSymbol: "m",
-                MultipleValues: false,
-                NearestValue: true,
-                Default: "500",
-                Current: null,
-                Extent: "500,490,480");
-        }
-
-        if (string.Equals(layer.Name, CiteAutosLayerTitle, StringComparison.OrdinalIgnoreCase))
-        {
-            return new CiteWmsDimensionDefinition(
-                Name: "time",
-                Units: "ISO8601",
-                UnitSymbol: null,
-                MultipleValues: true,
-                NearestValue: true,
-                Default: CiteAutosDefaultTime,
-                Current: null,
-                Extent: CiteAutosExtent);
-        }
-
-        return null;
-    }
-
-    private static async Task<string> BuildWmsCapabilities(
-        HttpContext context,
-        ServiceDefinition service,
-        string serviceId,
-        string baseUrl,
-        string version)
-    {
-        var isWms111 = IsWms111Version(version);
-        var responseVersion = isWms111 ? Wms111Version : Wms13Version;
-        var crsElementName = isWms111 ? "SRS" : "CRS";
-        var normalizedBaseUrl = baseUrl.TrimEnd('/');
-        var wmsEndpoint = $"{normalizedBaseUrl}/rest/services/{serviceId}/MapServer/WMS";
-        var wmsUrlPrefix = $"{wmsEndpoint}?";
-        var metadataUrl = $"{wmsEndpoint}?SERVICE=WMS&REQUEST=GetCapabilities&VERSION={responseVersion}";
-
-        var sb = new StringBuilder(8192);
-        sb.AppendLine("<?xml version=\"1.0\" encoding=\"UTF-8\"?>");
-        if (isWms111)
-        {
-            sb.Append("<!DOCTYPE WMT_MS_Capabilities SYSTEM \"")
-                .Append(Wms111CapabilitiesDtd)
-                .AppendLine("\">");
-            sb.Append("<WMT_MS_Capabilities ")
-                .Append("version=\"1.1.1\"")
-                .AppendLine(">");
-        }
-        else
-        {
-            sb.Append("<WMS_Capabilities ")
-                .Append("xmlns=\"http://www.opengis.net/wms\" ")
-                .Append("xmlns:xlink=\"http://www.w3.org/1999/xlink\" ")
-                .Append("xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" ")
-                .Append("version=\"1.3.0\" xsi:schemaLocation=\"")
-                .Append(WmsCapabilitiesSchemaLocation)
-                .AppendLine("\">");
-        }
-
-        sb.AppendLine("  <Service>");
-        sb.AppendLine("    <Name>WMS</Name>");
-        sb.Append("    <Title>").Append(EscapeXml(service.Name ?? serviceId)).AppendLine("</Title>");
-        sb.Append("    <Abstract>").Append(EscapeXml(service.Description ?? "Honua WMS service")).AppendLine("</Abstract>");
-        sb.AppendLine("    <KeywordList>");
-        sb.AppendLine("      <Keyword>WMS</Keyword>");
-        sb.AppendLine("      <Keyword>OGC</Keyword>");
-        sb.Append("      <Keyword>").Append(EscapeXml(service.Name ?? serviceId)).AppendLine("</Keyword>");
-        sb.AppendLine("    </KeywordList>");
-        AppendWmsOnlineResource(sb, "    ", wmsEndpoint, isWms111);
-        sb.AppendLine("    <ContactInformation>");
-        sb.AppendLine("      <ContactPersonPrimary>");
-        sb.AppendLine("        <ContactPerson>Honua Support</ContactPerson>");
-        sb.AppendLine("        <ContactOrganization>Honua</ContactOrganization>");
-        sb.AppendLine("      </ContactPersonPrimary>");
-        sb.AppendLine("      <ContactPosition>Support Engineer</ContactPosition>");
-        sb.AppendLine("      <ContactAddress>");
-        sb.AppendLine("        <AddressType>postal</AddressType>");
-        sb.AppendLine("        <Address>1 Honua Way</Address>");
-        sb.AppendLine("        <City>Honolulu</City>");
-        sb.AppendLine("        <StateOrProvince>HI</StateOrProvince>");
-        sb.AppendLine("        <PostCode>96815</PostCode>");
-        sb.AppendLine("        <Country>US</Country>");
-        sb.AppendLine("      </ContactAddress>");
-        sb.AppendLine("      <ContactVoiceTelephone>+1-555-0100</ContactVoiceTelephone>");
-        sb.AppendLine("      <ContactFacsimileTelephone>+1-555-0101</ContactFacsimileTelephone>");
-        sb.AppendLine("      <ContactElectronicMailAddress>support@honua.local</ContactElectronicMailAddress>");
-        sb.AppendLine("    </ContactInformation>");
-        sb.AppendLine("  </Service>");
-
-        sb.AppendLine("  <Capability>");
-        sb.AppendLine("    <Request>");
-
-        sb.AppendLine("      <GetCapabilities>");
-        sb.Append("        <Format>")
-            .Append(isWms111 ? Wms111CapabilitiesMimeType : WmsCapabilitiesMimeType)
-            .AppendLine("</Format>");
-        sb.AppendLine("        <DCPType>");
-        sb.AppendLine("          <HTTP>");
-        sb.AppendLine("            <Get>");
-        AppendWmsOnlineResource(sb, "              ", wmsUrlPrefix, isWms111);
-        sb.AppendLine("            </Get>");
-        sb.AppendLine("          </HTTP>");
-        sb.AppendLine("        </DCPType>");
-        sb.AppendLine("      </GetCapabilities>");
-
-        sb.AppendLine("      <GetMap>");
-        sb.AppendLine("        <Format>image/png</Format>");
-        sb.AppendLine("        <Format>image/jpeg</Format>");
-        sb.AppendLine("        <DCPType>");
-        sb.AppendLine("          <HTTP>");
-        sb.AppendLine("            <Get>");
-        AppendWmsOnlineResource(sb, "              ", wmsUrlPrefix, isWms111);
-        sb.AppendLine("            </Get>");
-        sb.AppendLine("          </HTTP>");
-        sb.AppendLine("        </DCPType>");
-        sb.AppendLine("      </GetMap>");
-
-        sb.AppendLine("      <GetFeatureInfo>");
-        sb.AppendLine("        <Format>text/plain</Format>");
-        sb.AppendLine("        <Format>application/json</Format>");
-        sb.AppendLine("        <DCPType>");
-        sb.AppendLine("          <HTTP>");
-        sb.AppendLine("            <Get>");
-        AppendWmsOnlineResource(sb, "              ", wmsUrlPrefix, isWms111);
-        sb.AppendLine("            </Get>");
-        sb.AppendLine("          </HTTP>");
-        sb.AppendLine("        </DCPType>");
-        sb.AppendLine("      </GetFeatureInfo>");
-
-        sb.AppendLine("    </Request>");
-        sb.AppendLine("    <Exception>");
-        if (isWms111)
-        {
-            sb.Append("      <Format>")
-                .Append(WmsSeXmlExceptionMimeType)
-                .AppendLine("</Format>");
-        }
-        else
-        {
-            sb.AppendLine("      <Format>XML</Format>");
-        }
-        sb.AppendLine("    </Exception>");
-
-        sb.AppendLine("    <Layer>");
-        sb.Append("      <Title>").Append(EscapeXml(service.Name ?? serviceId)).AppendLine("</Title>");
-        sb.Append("      <Abstract>").Append(EscapeXml(service.Description ?? "Honua WMS root layer")).AppendLine("</Abstract>");
-        sb.Append("      <").Append(crsElementName).AppendLine(">EPSG:4326</" + crsElementName + ">");
-        sb.Append("      <").Append(crsElementName).AppendLine(">EPSG:3857</" + crsElementName + ">");
-        if (!isWms111)
-        {
-            sb.AppendLine("      <CRS>CRS:84</CRS>");
-        }
-
-        if (service.EffectiveExtent.HasValue)
-        {
-            var rootExtent = service.EffectiveExtent.Value;
-            await AppendWmsGeographicBoundsAsync(context, sb, rootExtent, "      ", isWms111).ConfigureAwait(false);
-        }
-
-        var visibleLayers = service.Layers
-            .Where(l => l.HasGeometry && AccessPolicyHelpers.IsLayerAccessible(context, l, service))
-            .ToArray();
-        foreach (var layer in visibleLayers)
-        {
-            var layerName = GetWmsLayerName(layer);
-            var layerTitle = layer.Name ?? layerName;
-            var layerAbstract = layer.Description ?? $"WMS layer {layerName}";
-
-            sb.AppendLine("      <Layer queryable=\"1\">");
-            sb.Append("        <Name>").Append(EscapeXml(layerName)).AppendLine("</Name>");
-            sb.Append("        <Title>").Append(EscapeXml(layerTitle)).AppendLine("</Title>");
-            sb.Append("        <Abstract>").Append(EscapeXml(layerAbstract)).AppendLine("</Abstract>");
-            sb.AppendLine("        <KeywordList>");
-            sb.Append("          <Keyword>").Append(EscapeXml(layerName)).AppendLine("</Keyword>");
-            sb.AppendLine("        </KeywordList>");
-            sb.Append("        <").Append(crsElementName).AppendLine(">EPSG:4326</" + crsElementName + ">");
-            sb.Append("        <").Append(crsElementName).AppendLine(">EPSG:3857</" + crsElementName + ">");
-            if (!isWms111)
-            {
-                sb.AppendLine("        <CRS>CRS:84</CRS>");
-            }
-
-            var extent = layer.Extent ?? service.EffectiveExtent;
-            if (extent.HasValue)
-            {
-                await AppendWmsGeographicBoundsAsync(context, sb, extent.Value, "        ", isWms111).ConfigureAwait(false);
-            }
-
-            AppendWmsCiteDimensions(sb, layer, "        ", isWms111);
-            await AppendWmsTemporalDimensionAsync(context, sb, layer, "        ", isWms111).ConfigureAwait(false);
-
-            sb.AppendLine("        <MetadataURL type=\"TC211\">");
-            sb.AppendLine("          <Format>text/xml</Format>");
-            AppendWmsOnlineResource(sb, "          ", metadataUrl, isWms111);
-            sb.AppendLine("        </MetadataURL>");
-            sb.AppendLine("        <Style>");
-            sb.AppendLine("          <Name>default</Name>");
-            sb.AppendLine("          <Title>Default style</Title>");
-            sb.AppendLine("        </Style>");
-            sb.AppendLine("      </Layer>");
-        }
-
-        sb.AppendLine("    </Layer>");
-        sb.AppendLine("  </Capability>");
-        sb.Append(isWms111 ? "</WMT_MS_Capabilities>" : "</WMS_Capabilities>").AppendLine();
-        return sb.ToString();
-    }
-
-    private static async Task AppendWmsGeographicBoundsAsync(
-        HttpContext context,
-        StringBuilder sb,
-        FeatureExtent extent,
-        string indent,
-        bool isWms111)
-    {
-        var cancellationToken = TimeoutTokenHelper.GetTimeoutAwareCancellationToken(context);
-        var geographicExtent = extent;
-        if (extent.SpatialReference != 4326)
-        {
-            var transformResult = await TryTransformExtentAsync(
-                context,
-                new SkiaMapRenderer.RenderExtent(extent.MinX, extent.MinY, extent.MaxX, extent.MaxY),
-                extent.SpatialReference,
-                4326,
-                cancellationToken).ConfigureAwait(false);
-
-            if (!transformResult.IsSuccess)
-            {
-                return;
-            }
-
-            geographicExtent = FeatureExtent.Create(
-                transformResult.Extent.MinX,
-                transformResult.Extent.MinY,
-                transformResult.Extent.MaxX,
-                transformResult.Extent.MaxY,
-                4326);
-        }
-
-        if (isWms111)
-        {
-            AppendWms111LatLonBoundingBox(
-                sb,
-                geographicExtent.MinX,
-                geographicExtent.MinY,
-                geographicExtent.MaxX,
-                geographicExtent.MaxY,
-                indent);
-            AppendWmsBoundingBox(
-                sb,
-                "EPSG:4326",
-                geographicExtent.MinX,
-                geographicExtent.MinY,
-                geographicExtent.MaxX,
-                geographicExtent.MaxY,
-                indent,
-                isWms111);
-        }
-        else
-        {
-            AppendWmsGeographicBoundingBox(
-                sb,
-                geographicExtent.MinX,
-                geographicExtent.MinY,
-                geographicExtent.MaxX,
-                geographicExtent.MaxY,
-                indent);
-            AppendWmsBoundingBox(
-                sb,
-                "CRS:84",
-                geographicExtent.MinX,
-                geographicExtent.MinY,
-                geographicExtent.MaxX,
-                geographicExtent.MaxY,
-                indent,
-                isWms111);
-            AppendWmsBoundingBox(
-                sb,
-                "EPSG:4326",
-                geographicExtent.MinX,
-                geographicExtent.MinY,
-                geographicExtent.MaxX,
-                geographicExtent.MaxY,
-                indent,
-                isWms111);
-        }
-    }
-
-    private static void AppendWmsGeographicBoundingBox(
-        StringBuilder sb,
-        double minX,
-        double minY,
-        double maxX,
-        double maxY,
-        string indent)
-    {
-        sb.Append(indent).AppendLine("<EX_GeographicBoundingBox>");
-        sb.Append(indent).Append("  <westBoundLongitude>").Append(minX.ToString("F6", CultureInfo.InvariantCulture)).AppendLine("</westBoundLongitude>");
-        sb.Append(indent).Append("  <eastBoundLongitude>").Append(maxX.ToString("F6", CultureInfo.InvariantCulture)).AppendLine("</eastBoundLongitude>");
-        sb.Append(indent).Append("  <southBoundLatitude>").Append(minY.ToString("F6", CultureInfo.InvariantCulture)).AppendLine("</southBoundLatitude>");
-        sb.Append(indent).Append("  <northBoundLatitude>").Append(maxY.ToString("F6", CultureInfo.InvariantCulture)).AppendLine("</northBoundLatitude>");
-        sb.Append(indent).AppendLine("</EX_GeographicBoundingBox>");
-    }
-
-    private static void AppendWmsBoundingBox(
-        StringBuilder sb,
-        string crs,
-        double minX,
-        double minY,
-        double maxX,
-        double maxY,
-        string indent,
-        bool isWms111)
-    {
-        var outputMinX = minX;
-        var outputMinY = minY;
-        var outputMaxX = maxX;
-        var outputMaxY = maxY;
-        if (!isWms111 &&
-            string.Equals(crs, "EPSG:4326", StringComparison.OrdinalIgnoreCase))
-        {
-            outputMinX = minY;
-            outputMinY = minX;
-            outputMaxX = maxY;
-            outputMaxY = maxX;
-        }
-
-        sb.Append(indent)
-            .Append("<BoundingBox ")
-            .Append(isWms111 ? "SRS" : "CRS")
-            .Append("=\"")
-            .Append(EscapeXml(crs))
-            .Append("\" minx=\"")
-            .Append(outputMinX.ToString("F6", CultureInfo.InvariantCulture))
-            .Append("\" miny=\"")
-            .Append(outputMinY.ToString("F6", CultureInfo.InvariantCulture))
-            .Append("\" maxx=\"")
-            .Append(outputMaxX.ToString("F6", CultureInfo.InvariantCulture))
-            .Append("\" maxy=\"")
-            .Append(outputMaxY.ToString("F6", CultureInfo.InvariantCulture))
-            .AppendLine("\" />");
-    }
-
-    private static void AppendWms111LatLonBoundingBox(
-        StringBuilder sb,
-        double minX,
-        double minY,
-        double maxX,
-        double maxY,
-        string indent)
-    {
-        sb.Append(indent)
-            .Append("<LatLonBoundingBox minx=\"")
-            .Append(minX.ToString("F6", CultureInfo.InvariantCulture))
-            .Append("\" miny=\"")
-            .Append(minY.ToString("F6", CultureInfo.InvariantCulture))
-            .Append("\" maxx=\"")
-            .Append(maxX.ToString("F6", CultureInfo.InvariantCulture))
-            .Append("\" maxy=\"")
-            .Append(maxY.ToString("F6", CultureInfo.InvariantCulture))
-            .AppendLine("\" />");
-    }
-
     private enum CiteTerrainSelection
     {
         DefaultRange,
@@ -2622,16 +1543,6 @@ internal static class WmsRequestHandlers
         LowRangeAndValue250,
         Value250
     }
-
-    private readonly record struct CiteWmsDimensionDefinition(
-        string Name,
-        string Units,
-        string? UnitSymbol,
-        bool MultipleValues,
-        bool NearestValue,
-        string? Default,
-        string? Current,
-        string Extent);
 
     private sealed class WmsImageResult : IResult
     {
